update prometheus.yml and move serviceb package
