+++ Tag 2/05_restClient/aufgabe/prometheus.yml
@@ -7,3 +7,8 @@ scrape_configs:
       - targets: ['host.docker.internal:8080']
     metrics_path: '/q/metrics'
     scrape_interval: 1s
+  - job_name: service_b
+    static_configs:
+      - targets: ['host.docker.internal:8082']
+    metrics_path: '/q/metrics'
+    scrape_interval: 1s
