+++ Tag 2/05_restClient/aufgabe/service_b/src/main/java/qs/mp/serviceb/MyLivenessCheck.java
@@ -1,4 +1,4 @@
-package qs.mp;
+package qs.mp.serviceb;
 
 import org.eclipse.microprofile.health.HealthCheck;
 import org.eclipse.microprofile.health.HealthCheckResponse;
