+++ Tag 2/05_restClient/aufgabe/service_b/src/main/java/qs/mp/serviceb/control/MessageService.java
@@ -1,18 +1,14 @@
-package qs.mp.control;
+package qs.mp.serviceb.control;
 
 import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-import org.eclipse.microprofile.config.ConfigProvider;
-import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.eclipse.microprofile.metrics.annotation.Counted;
 import org.eclipse.microprofile.metrics.annotation.Gauge;
 import org.eclipse.microprofile.metrics.annotation.Timed;
-import qs.mp.entity.Author;
-import qs.mp.entity.Message;
+import qs.mp.serviceb.entity.Author;
+import qs.mp.serviceb.entity.Message;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 
 @ApplicationScoped
 @Timed
