+++ Tag 2/05_restClient/aufgabe/service_b/src/main/java/qs/mp/serviceb/entity/Author.java
@@ -1,4 +1,4 @@
-package qs.mp.entity;
+package qs.mp.serviceb.entity;
 
 import lombok.Data;
