+++ Tag 2/05_restClient/aufgabe/service_b/src/main/java/qs/mp/serviceb/entity/Message.java
@@ -1,4 +1,4 @@
-package qs.mp.entity;
+package qs.mp.serviceb.entity;
 
 import lombok.Data;
 
@@ -7,9 +7,9 @@ public class Message {
 
     private String heading;
     private String body;
-    private Author author;
+    private qs.mp.serviceb.entity.Author author;
 
-    public Message(String heading, String body, Author author) {
+    public Message(String heading, String body, qs.mp.serviceb.entity.Author author) {
         this.heading = heading;
         this.body = body;
         this.author = author;
