+++ Tag 2/05_restClient/aufgabe/service_b/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+quarkus.http.port=8082
